migrate helper stamp functions
